==> 3_features_analysis.ipynb <==
# Features Analysis

This notebook explore the features of the dataset and their relationship with the target variable. It aims to understand and generate new features that can be used to improve the model. Those features will be tested by a very simple random forest model to check their relevance in this classification task.

Inspired from https://www.kaggle.com/code/carlmcbrideellis/zzzs-random-forest-model-starter/notebook

import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
from itertools import groupby
import gc
import polars as pl
import seaborn as sns
from sklearn.model_selection import train_test_split
import pickle
from sklearn.metrics import classification_report, confusion_matrix, accuracy_score
from sklearn.ensemble import RandomForestClassifier


df_signals = pl.read_parquet("data/train_series.parquet")
df_events = pl.read_csv("data/train_events.csv")

df_combined = df_signals.join(df_events, on=['series_id', 'timestamp'], how='inner')

with open("data//df_combined.pkl", "rb") as f:
   pickle.dump(df_combined, f)

with open("data/df_combined.pkl", "rb") as f:
   df_combined = pickle.load(f)

### Creation of new features based on z-angle

Possible features to be created from the z-angle:
- **diff_z_angle**: difference between the z-angle between two consecutive frames
- **z_angle_mean_window**: mean of the z-angle in a period of time
- **z_angle_std_window**: standard deviation of the z-angle in a period of time
- **z_angle_min_window**: minimum value of the z-angle in a period of time
- **z_angle_max_window**: maximum value of the z-angle in a period of time
with the window being a period of 2 to 50 frames

def make_features_anglez(df, periods):
   """
   Compute features for angle_z

   Param:
   df: dataframe
   periods: list of periods to compute the features

   Return:
   df: dataframe with the new features
   """
   for p in periods:
      # for each series in the dataset
      for serie in df['series_id'].unique():
         df.loc[df['series_id'] == serie, 'angle_z_mean_' + str(p)] = df.loc[df['series_id'] == serie, 'angle_z'].rolling(window=p).mean()
         

def make_features_enmo(df, periods):
   for p in periods:
      df[f'enmo_diff_{p}'] = df['enmo'].diff(periods=p).fillna(method="bfill").astype('float16')
      df[f'enmo_rolling_mean_{p}'] = df['enmo'].rolling(p).mean().fillna(method="bfill").astype('float16')
      df[f'enmo_rolling_std_{p}'] = df['enmo'].rolling(p).std().fillna(method="bfill").astype('float16')
      df[f'enmo_rolling_min_{p}'] = df['enmo'].rolling(p).min().fillna(method="bfill").astype('float16')
      df[f'enmo_rolling_max_{p}'] = df['enmo'].rolling(p).max().fillna(method="bfill").astype('float16')
      df[f'enmo_rolling_median_{p}'] = df['enmo'].rolling(p).median().fillna(method="bfill").astype('float16')
      df[f'enmo_rolling_skew_{p}'] = df['enmo'].rolling(p).skew().fillna(method="bfill").astype('float16')
                     

def make_features(df, periods=20):
   # transform angle_z values by their absolute value
   df['anglez'] = df['anglez'].abs()
   df['timestamp'].to_datetime()
   df = make_features_anglez(df, periods)
   df = make_features_enmo(df, periods)
   return df

df.head()

periods = [10,20,50]
df = make_features(df, periods)

features = [col for col in df.columns if col not in ['enmo']]
features

X = df[features]
y = df["event"]

X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.2, random_state=42)

# save some memory
del X
del y
gc.collect();

classifier = RandomForestClassifier(random_state=42,n_jobs=-1)

classifier.fit(X_train, y_train)

# save some memory
del X_train, y_train
gc.collect();

predict = classifier.predict(X_test)

print(classification_report(y_test,predict))
print(confusion_matrix(y_test,predict))
print(accuracy_score(y_test, predict))
