==> 4_feature_selection.ipynb <==
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns
import polars as pl
from tsfresh import extract_features
from tsfresh import select_features

from tqdm import tqdm

df = pl.read_parquet('data/train_dataset_feature.parquet').to_pandas()

df.shape

X = df.drop(['series_id', 'timestamp','onset','wakeup','state'], axis=1)
y = df['state']
onset = df['onset']
wakeup = df['wakeup']
series_id = df['series_id']

features_filtered = select_features(X, y)


features_filtered.shape

features_filtered


X = pl.from_pandas(features_filtered)
y = pl.from_pandas(y).to_frame()
onset = pl.from_pandas(onset).to_frame()
wakeup = pl.from_pandas(wakeup).to_frame()
series_id = pl.from_pandas(series_id).to_frame()

#Concatenate the features and the labels

df = pl.concat([X, y,onset,wakeup,series_id], how="horizontal")

df.write_parquet('data/train_dataset_feature_selected.parquet')
